requirements.txt
streamlit
openai
wikipedia
pydub
pypdf
pillow
python-docx